fix: multi-platform integration audit (UE5, Unity, Godot)
UE5:
- Fix alice_sdf_mirror ABI mismatch (float → uint8_t)
- Fix alice_sdf_repeat_finite signedness (int32_t → uint32_t)
- Remove invalid BlueprintSpawnableComponent on AActor
- Add BatchResult error check in EvalDistanceBatch
- Remove unused ALICE_BINARY_OP macro
- Update plugin version to 1.1.0

Unity:
- Fix P/Invoke mirror (float → byte) and repeat_finite (int → uint)
- Add SdfNodeWrapper.cs: AliceSdf namespace with SdfNode, CompiledSdf, Native

Godot:
- Fix library paths to match README (add /bin/ subdirectory)
- Add macOS ARM64 architecture keys
- Fix get_vertex_count returning surface count instead of vertex count
- Fix get_lod_vertex_counts returning surface count instead of vertex count

+++ unity-sdf-universe/Assets/Scripts/AliceSdf.cs
@@ -527,8 +527,8 @@ namespace AliceSdfUnity
         public static IntPtr Repeat(IntPtr node, Vector3 spacing) =>
             alice_sdf_repeat(node, spacing.x, spacing.y, spacing.z);
 
-        [DllImport(LibraryName)] private static extern IntPtr alice_sdf_repeat_finite(IntPtr node, int cx, int cy, int cz, float sx, float sy, float sz);
-        [DllImport(LibraryName)] private static extern IntPtr alice_sdf_mirror(IntPtr node, float mx, float my, float mz);
+        [DllImport(LibraryName)] private static extern IntPtr alice_sdf_repeat_finite(IntPtr node, uint cx, uint cy, uint cz, float sx, float sy, float sz);
+        [DllImport(LibraryName)] private static extern IntPtr alice_sdf_mirror(IntPtr node, byte mx, byte my, byte mz);
         [DllImport(LibraryName)] private static extern IntPtr alice_sdf_elongate(IntPtr node, float ex, float ey, float ez);
         [DllImport(LibraryName)] private static extern IntPtr alice_sdf_revolution(IntPtr node, float offset);
         [DllImport(LibraryName)] private static extern IntPtr alice_sdf_extrude(IntPtr node, float half_height);
@@ -546,9 +546,9 @@ namespace AliceSdfUnity
         public static IntPtr Animated(IntPtr node, float speed, float amplitude) => alice_sdf_animated(node, speed, amplitude);
 
         public static IntPtr RepeatFinite(IntPtr node, Vector3Int count, Vector3 spacing) =>
-            alice_sdf_repeat_finite(node, count.x, count.y, count.z, spacing.x, spacing.y, spacing.z);
+            alice_sdf_repeat_finite(node, (uint)count.x, (uint)count.y, (uint)count.z, spacing.x, spacing.y, spacing.z);
         public static IntPtr Mirror(IntPtr node, bool x, bool y, bool z) =>
-            alice_sdf_mirror(node, x ? 1f : 0f, y ? 1f : 0f, z ? 1f : 0f);
+            alice_sdf_mirror(node, (byte)(x ? 1 : 0), (byte)(y ? 1 : 0), (byte)(z ? 1 : 0));
         public static IntPtr Elongate(IntPtr node, Vector3 amount) =>
             alice_sdf_elongate(node, amount.x, amount.y, amount.z);
         public static IntPtr Revolution(IntPtr node, float offset) => alice_sdf_revolution(node, offset);
